backend/requirements.txt
# Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Banco
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth & Segurança
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# Validação
pydantic==2.7.0
pydantic-settings==2.2.1
email-validator==2.1.1

# Rate limiting
slowapi==0.1.9

# HTTP client (para verificar tokens Google)
httpx==0.27.0

# Logs estruturados
structlog==24.1.0

# Testes
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
pytest-cov==5.0.0
factory-boy==3.3.0
